[C8] Split sweep_loft.go into sweep + loft (#2240)
model/feature/sweep_loft.go was 863 LOC, 1.7x the 500-line ceiling — two
distinct features sharing only the profile/path binding helpers. Split into
four files, all under the ceiling:

- sweep.go (205): the sweep feature (definition, Recompute, orientation, the
  swept-section tool builder) and the SweepFeatures collection.
- loft.go (321): the loft feature (types, definition, Recompute, end/guide
  conditions) and the LoftFeatures collection.
- loft_sections.go (287): the loft section resolution (sketch loop / point /
  model face) and the skin/tube helpers that build the loft tool body.
- sweep_loft.go (98): the shared binding helpers (centroidOf, resampleLoop,
  edgeInteriorCounts, loopSegmentLengths).

Behaviour-preserving: split programmatically by receiver-qualified
declaration, verified every non-comment code line is byte-identical (only
import redistribution differs); imports resolved per file with goimports. The
moved functions carry no funlen/gocyclo issue, so no .golangci.yml change is
needed (the remaining sweep_loft.go still matches the existing exclusion).
Feature package builds, archguard passes, lint 0. Net delta of recognizers /
tolerance constants / fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2240.

(cherry picked from commit 535dfc52e46fb514162213f1e8ed86e54eb10316)
